Move the UUID extension dtype into encodings/uuid
`vortex.uuid` was split across `vortex-array` (the extension dtype and its
vtable) and `vortex-arrow` (the Arrow canonical `arrow.uuid` import/export
plugins), and both crates registered it unconditionally in their default
sessions. Pull both halves into a new `vortex-uuid` crate under `encodings/`,
following the same shape as `vortex-parquet-variant`: the dtype, its Arrow
plugins, and an `initialize` that registers all of them with a session.

The default `VortexSession` calls `vortex_uuid::initialize`, so sessions built
through the `vortex` facade are unchanged. Sessions built from
`vortex_array::array_session()` alone no longer resolve `vortex.uuid` and must
initialize the crate themselves.

Supporting changes:

- `vortex_arrow::nulls` and `vortex_arrow::has_valid_extension_type` are now
  public, since the UUID Arrow plugins live outside the crate.
- `ArrowSession` derives `Default` now that it pre-registers no plugins.
- The UUID-based tests in `vortex-arrow` (which used UUID as their exemplar
  Arrow extension type) move to `vortex-uuid` alongside the plugins.
- Downstream users import from `vortex::encodings::uuid` / `vortex_uuid`.
