+++ encodings/parquet-variant/Cargo.toml
@@ -41,3 +41,4 @@ vortex-array = { workspace = true, features = ["_test-harness"] }
 vortex-file = { workspace = true, features = ["tokio"] }
 vortex-io = { workspace = true, features = ["tokio"] }
 vortex-layout = { workspace = true }
+vortex-uuid = { workspace = true }
