+++ vortex-array/src/dtype/session.rs
@@ -17,7 +17,6 @@ use crate::dtype::extension::ExtVTable;
 use crate::extension::datetime::Date;
 use crate::extension::datetime::Time;
 use crate::extension::datetime::Timestamp;
-use crate::extension::uuid::Uuid;
 
 /// Registry for extension dtypes.
 pub type ExtDTypeRegistry = ArcSwapMap<Id, ExtDTypePluginRef>;
@@ -38,7 +37,6 @@ impl Default for DTypeSession {
         this.register(Date);
         this.register(Time);
         this.register(Timestamp);
-        this.register(Uuid);
 
         this
     }
