+++ vortex-duckdb/src/convert/scalar.rs
@@ -28,6 +28,7 @@ use vortex::dtype::PType;
 use vortex::dtype::PType::I32;
 use vortex::dtype::PType::I64;
 use vortex::dtype::half::f16;
+use vortex::encodings::uuid::Uuid;
 use vortex::error::VortexError;
 use vortex::error::VortexResult;
 use vortex::error::vortex_bail;
@@ -40,7 +41,6 @@ use vortex::extension::datetime::Time;
 use vortex::extension::datetime::TimeUnit;
 use vortex::extension::datetime::Timestamp;
 use vortex::extension::datetime::TimestampOptions;
-use vortex::extension::uuid::Uuid;
 use vortex::scalar::BinaryScalar;
 use vortex::scalar::BoolScalar;
 use vortex::scalar::DecimalScalar;
