+++ vortex-duckdb/src/exporter/extension.rs
@@ -6,9 +6,9 @@ use vortex::array::arrays::ExtensionArray;
 use vortex::array::arrays::TemporalArray;
 use vortex::array::arrays::extension::ExtensionArrayExt;
 use vortex::array::extension::datetime::AnyTemporal;
+use vortex::encodings::uuid::Uuid;
 use vortex::error::VortexResult;
 use vortex::error::vortex_bail;
-use vortex::extension::uuid::Uuid;
 use vortex_spatial::extension::LineString;
 use vortex_spatial::extension::LineStringData;
 use vortex_spatial::extension::MultiLineString;
