+++ vortex/Cargo.toml
@@ -52,6 +52,7 @@ vortex-session = { workspace = true }
 vortex-sparse = { workspace = true }
 vortex-tensor = { workspace = true, optional = true }
 vortex-utils = { workspace = true }
+vortex-uuid = { workspace = true }
 vortex-zigzag = { workspace = true }
 vortex-zstd = { workspace = true, optional = true }
